fix transfer delete

+++ pkg/vm/engine/disttae/logtail.go
@@ -66,19 +66,16 @@ func consumeEntry(
 		switch e.TableId {
 		case catalog.MO_TABLES_ID:
 			bat, _ := batch.ProtoBatchToBatch(e.Bat)
-			logutil.Infof("yyyyy handle tbl table batch %d", bat.RowCount())
 			if cache != nil {
 				cache.InsertTable(bat)
 			}
 		case catalog.MO_DATABASE_ID:
 			bat, _ := batch.ProtoBatchToBatch(e.Bat)
-			logutil.Infof("yyyyy handle db table batch %d", bat.RowCount())
 			if cache != nil {
 				cache.InsertDatabase(bat)
 			}
 		case catalog.MO_COLUMNS_ID:
 			bat, _ := batch.ProtoBatchToBatch(e.Bat)
-			logutil.Infof("yyyyy handle col table batch %d", bat.RowCount())
 			if cache != nil {
 				cache.InsertColumns(bat)
 			}
@@ -89,13 +86,13 @@ func consumeEntry(
 
 	switch e.TableId {
 	case catalog.MO_TABLES_ID:
-		bat, _ := batch.ProtoBatchToBatch(e.Bat)
-		if cache != nil {
+		if cache != nil && !logtailreplay.IsTransferredDels(e.TableName) {
+			bat, _ := batch.ProtoBatchToBatch(e.Bat)
 			cache.DeleteTable(bat)
 		}
 	case catalog.MO_DATABASE_ID:
-		bat, _ := batch.ProtoBatchToBatch(e.Bat)
-		if cache != nil {
+		if cache != nil && !logtailreplay.IsTransferredDels(e.TableName) {
+			bat, _ := batch.ProtoBatchToBatch(e.Bat)
 			cache.DeleteDatabase(bat)
 		}
 	}
